Add list process function and error handling
The new process function helps glue together the many utility functions
that produce the desired data structure.

Did some error handling in case I need to change some of the default
arguments. This is in hopes of reusing the functions for parsing
the other output files I have left to convert.

The diff will look kind of crazy because I reorganized the code a bit.