Rework file methods, fix bug in 'file_to_lines' method
Make file methods work as they should. Some arguments were not working.
Also I was still trying to figure out what the methods should be doing.

The 'file_to_lines' method was only producing one dictionary value
to represent files in the server directory rather than all of them.
This is because each time through the loop it would overwrite to same key
value. initializing the varibale to a empty dict and using the 'update'
dict method fixed the behavior.

Minor change to the file path variable as well